Refactor LLM service: consolidate prompt generation and summary functions into llm.py; remove llm_service.py; update tests accordingly
